feat: sql coalesce() support
